Make linux check for subprocess errors
MacOS, Windows and WSL wait for the open child process to exit, and
checks its exit code. This means that errors in the opening process
are returned to the user. However, the linux implementation does not
wait for the xdg-open process to exit. This means that potential
errors returned by xdg-open are ignored.

This fixes the code by making the linux implementation also wait for
the child process to exit, and checking its return type.