Replace winapi dependency with windows
Since normpath was [added][1], opener has depended on both winapi (directly)
and windows-sys (transitively).

This commit replaces usage of winapi with the windows crate so that it
only depends on a single windows API crate. Additionally, the windows
crate includes some functions that were manually linked in opener, so
those manual links were able to be cleaned up.

[1]: 3b48ee084300ba883bed11259d978fa989ffaed8